- Updated UI test base;

+++ MeMe/app/src/main/java/com/spandr/meme/core/activity/main/logic/builder/draggable/common/data/AbstractDataProvider.java
@@ -6,7 +6,7 @@ import android.view.View;
 public abstract class AbstractDataProvider {
 
     public static abstract class Data {
-        public abstract long getId();
+        public abstract int getId();
 
         public abstract boolean isSectionHeader();
